Add abstract class Operation2 to reduce code duplication
