+++ h2/src/main/org/h2/expression/CompatibilityDatePlusTimeOperation.java
@@ -9,8 +9,6 @@ import static org.h2.util.DateTimeUtils.NANOS_PER_DAY;
 
 import org.h2.engine.Session;
 import org.h2.message.DbException;
-import org.h2.table.ColumnResolver;
-import org.h2.table.TableFilter;
 import org.h2.util.DateTimeUtils;
 import org.h2.value.DataType;
 import org.h2.value.TypeInfo;
@@ -25,15 +23,10 @@ import org.h2.value.ValueTimestampTimeZone;
 /**
  * A compatibility mathematical operation with datetime values.
  */
-public class CompatibilityDatePlusTimeOperation extends Expression {
-
-    private Expression left, right;
-
-    private TypeInfo type;
+public class CompatibilityDatePlusTimeOperation extends Operation2 {
 
     public CompatibilityDatePlusTimeOperation(Expression left, Expression right) {
-        this.left = left;
-        this.right = right;
+        super(left, right);
         TypeInfo l = left.getType(), r = right.getType();
         int t;
         switch (l.getValueType()) {
@@ -107,12 +100,6 @@ public class CompatibilityDatePlusTimeOperation extends Expression {
     }
 
     @Override
-    public void mapColumns(ColumnResolver resolver, int level, int state) {
-        left.mapColumns(resolver, level, state);
-        right.mapColumns(resolver, level, state);
-    }
-
-    @Override
     public Expression optimize(Session session) {
         left = left.optimize(session);
         right = right.optimize(session);
@@ -120,50 +107,6 @@ public class CompatibilityDatePlusTimeOperation extends Expression {
             return ValueExpression.get(getValue(session));
         }
         return this;
-    }
-
-    @Override
-    public void setEvaluatable(TableFilter tableFilter, boolean b) {
-        left.setEvaluatable(tableFilter, b);
-        right.setEvaluatable(tableFilter, b);
-    }
-
-    @Override
-    public TypeInfo getType() {
-        return type;
-    }
-
-    @Override
-    public void updateAggregate(Session session, int stage) {
-        left.updateAggregate(session, stage);
-        right.updateAggregate(session, stage);
-    }
-
-    @Override
-    public boolean isEverything(ExpressionVisitor visitor) {
-        return left.isEverything(visitor) && right.isEverything(visitor);
-    }
-
-    @Override
-    public int getCost() {
-        return left.getCost() + 1 + right.getCost();
-    }
-
-    @Override
-    public int getSubexpressionCount() {
-        return 2;
-    }
-
-    @Override
-    public Expression getSubexpression(int index) {
-        switch (index) {
-        case 0:
-            return left;
-        case 1:
-            return right;
-        default:
-            throw new IndexOutOfBoundsException();
-        }
     }
 
 }
