+++ h2/src/main/org/h2/expression/Operation2.java
@@ -1,0 +1,77 @@
+/*
+ * Copyright 2004-2020 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * and the EPL 1.0 (https://h2database.com/html/license.html).
+ * Initial Developer: H2 Group
+ */
+package org.h2.expression;
+
+import org.h2.engine.Session;
+import org.h2.table.ColumnResolver;
+import org.h2.table.TableFilter;
+import org.h2.value.TypeInfo;
+
+/**
+ * Operation with two arguments.
+ */
+abstract class Operation2 extends Expression {
+
+    Expression left, right;
+
+    TypeInfo type;
+
+    protected Operation2(Expression left, Expression right) {
+        this.left = left;
+        this.right = right;
+    }
+
+    @Override
+    public TypeInfo getType() {
+        return type;
+    }
+
+    @Override
+    public void mapColumns(ColumnResolver resolver, int level, int state) {
+        left.mapColumns(resolver, level, state);
+        right.mapColumns(resolver, level, state);
+    }
+
+    @Override
+    public void setEvaluatable(TableFilter tableFilter, boolean value) {
+        left.setEvaluatable(tableFilter, value);
+        right.setEvaluatable(tableFilter, value);
+    }
+
+    @Override
+    public void updateAggregate(Session session, int stage) {
+        left.updateAggregate(session, stage);
+        right.updateAggregate(session, stage);
+    }
+
+    @Override
+    public boolean isEverything(ExpressionVisitor visitor) {
+        return left.isEverything(visitor) && right.isEverything(visitor);
+    }
+
+    @Override
+    public int getCost() {
+        return left.getCost() + right.getCost() + 1;
+    }
+
+    @Override
+    public int getSubexpressionCount() {
+        return 2;
+    }
+
+    @Override
+    public Expression getSubexpression(int index) {
+        switch (index) {
+        case 0:
+            return left;
+        case 1:
+            return right;
+        default:
+            throw new IndexOutOfBoundsException();
+        }
+    }
+
+}
